requirements.txt
Flask==2.0.1
requests==2.26.0
Werkzeug==2.1.0
psycopg2-binary==2.9.1
